LayoutTests/animations/trigger-container-scroll-boundaries-expected.txt

This element should begin animating only when the page scrolls to 20px from the top. It then animates smoothly, moving 100px to the left over the next 100px of scrolling. Remember to scroll to the top of the page before reloading!

PASS: Value before animation is applied is 0px
PASS: Value with animation but no scroll was 0px
FAIL: Value with scroll amount of 10 should be 10px, was 100px
FAIL: Value with scroll amount of 20 should be 20px, was 100px
FAIL: Value with scroll amount of 30 should be 30px, was 100px
FAIL: Value with scroll amount of 40 should be 40px, was 100px
FAIL: Value with scroll amount of 50 should be 50px, was 100px
FAIL: Value with scroll amount of 60 should be 60px, was 100px
FAIL: Value with scroll amount of 70 should be 70px, was 100px
FAIL: Value with scroll amount of 80 should be 80px, was 100px
FAIL: Value with scroll amount of 90 should be 90px, was 100px
